// frontend/src/app/shared/generated-api/model/userSettings.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserSettings { 
    theme: UserSettingsThemeEnum;
    language: UserSettingsLanguageEnum;
    moveCheckedToBottom: boolean;
    trashAutoPurgeDays: number;
    trashAutoPurgeMayBeDisabled: boolean;
}
export enum UserSettingsThemeEnum {
    Dark = 'dark',
    Light = 'light'
};
export enum UserSettingsLanguageEnum {
    En = 'en',
    De = 'de'
};